+++ app/app.js
@@ -1,16 +1,9 @@
-angular.module('app').config(function($locationProvider, $stateProvider, $urlRouterProvider) {
-  // Use HTML5 push history
+'use strict';
+
+angular.module('app').config(function($locationProvider, $routeProvider) {
   $locationProvider.html5Mode(true);
 
-  // Set up routes
-  $stateProvider.state('index', {
-    url: '/',
-    templateUrl: 'templates/index.html'
-  });
-
-  // Set up default route
-  $urlRouterProvider.otherwise('/');
-})
-.run(function($rootScope) {
-  $rootScope.title = 'Aang Brunch';
+  $routeProvider
+  .when('/', {templateUrl: 'templates/index.html'})
+  .otherwise('/');
 });