+++ brunch-config.js
@@ -1,20 +1,20 @@
-var setup = require('./setup');
+'use strict';
 
-exports.config = setup({
-  modules: {
-    definition: false,
-    wrapper: false
+exports.config = {
+  paths: {
+    public: 'public'
   },
 
   server: {
-    path: 'server'
+    path: 'server',
+    port: 3333
   },
 
   files: {
     javascripts: {
       joinTo: {
         'javascripts/app.js': /^app/,
-        'javascripts/vendor.js': /^(vendor|bower_components)/
+        'javascripts/vendor.js': /^(vendor|bower_components)(?!.+angular-mocks.js$)/
       },
       order: {
         before: ['app/config.js']
@@ -30,5 +30,15 @@ exports.config = setup({
     templates: {
       joinTo: 'javascripts/app.js'
     }
+  },
+
+  modules: {
+    definition: false,
+    wrapper: function(path, code) {
+      if(/^app\//.test(path)) {
+        code = '\n(function() {\n' + code + '\n})();';
+      }
+      return code;
+    }
   }
-});
+};