+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+### 0.5.0 (February 10, 2014)
+- Enforce strict mode
+- Go back to ngRoute for now (too many issues with ui-router)
+- Minor refactor in skeleton
+- Update BTC
+  - Remove Prerender server
+  - Add support for Mongoose/Passport/sessions
+  - Simplify Brunch config
+  - Remove dashes in generators
+  - Detect browsers in Karma
+  - Server refactoring
+  - Unify `add:codetesting` and `add:sitetesting` tasks
+  - Add scaffold to generate server model
+  - Add aliases to generator tasks
+
 ### 0.4.0 (February 5, 2014)
 - Add angular-mocks to code testing package set
 - Update BTC